With `Lexer` token `Span` positioning and do cleanup

+++ include/parser/Lexer.h
@@ -29,7 +29,7 @@ namespace jc::parser {
         token_list tokens;
 
         // Lexer current position
-        uint64_t index{0};
+        uint32_t index{0};
         Location loc;
         std::vector<uint32_t> linesIndices;
